+++ crates/bevy_tasks/examples/busy_behavior.rs
@@ -8,31 +8,27 @@ use web_time::{Duration, Instant};
 fn main() {
     let pool = TaskPoolBuilder::new()
         .thread_name("Busy Behavior ThreadPool".to_string())
-        .num_threads(5)
+        .num_threads(4)
         .build();
 
-    println!("main {:?}", std::thread::current().id());
-    const iter_count: usize = 1000;
     let t0 = Instant::now();
-    let a: Vec<_> = (0..20000).collect();
-    for _ in 0..iter_count {
-        pool.scope(|s| {
-            for i in 0..20 {
-                s.spawn(async move {
-                    let now = Instant::now();
-                    while Instant::now() - now < Duration::from_micros(1000) {
-                        // spin, simulating work being done
-                    }
+    pool.scope(|s| {
+        for i in 0..40 {
+            s.spawn(async move {
+                let now = Instant::now();
+                while Instant::now() - now < Duration::from_millis(100) {
+                    // spin, simulating work being done
+                }
 
-                    // println!(
-                    //     "Thread {:?} index {} finished",
-                    //     std::thread::current().id(),
-                    //     i
-                    // );
-                });
-            }
-        });
-    }
+                println!(
+                    "Thread {:?} index {} finished",
+                    std::thread::current().id(),
+                    i
+                );
+            });
+        }
+    });
 
-    println!(" par {:?} elapsed", t0.elapsed() / iter_count as u32);
+    let t1 = Instant::now();
+    println!("all tasks finished in {} secs", (t1 - t0).as_secs_f32());
 }
