+++ crates/bevy_tasks/src/usages.rs
@@ -74,11 +74,11 @@ taskpool! {
     (IO_TASK_POOL, IoTaskPool)
 }
 
-
+/// get thread numbers of bevy's compute task pool
 pub fn compute_task_pool_thread_num() -> usize {
     ComputeTaskPool::try_get()
         .map(|p| p.thread_num())
-        .unwrap_or(1)
+        .unwrap_or(0)
 }
 
 /// A function used by `bevy_core` to tick the global tasks pools on the main thread.
